teams-agent/requirements.txt
azure-functions>=1.19.0
azure-storage-blob>=12.19.0
azure-data-tables>=12.4.0
azure-servicebus>=7.11.0
botbuilder-core>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
botbuilder-schema>=4.15.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
